dom/plugins/test/mochitest/test_bug751809.html

<html>
<head>
  <title>Bug 751809</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
  <script class="testbody" type="application/javascript">
  
  SimpleTest.waitForExplicitFinish();
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  const Ci = Components.interfaces;
  const clickToPlayPref = "plugins.click_to_play";
  const utils = window.QueryInterface(Ci.nsIInterfaceRequestor).
                                    getInterface(Ci.nsIDOMWindowUtils);

  function waitForCondition(condition, nextTest, errorMsg) {
    var tries = 0;
    var interval = setInterval(function() {
      if (tries >= 30) {
        ok(false, errorMsg);
        moveOn();
      }
      if (condition()) {
        moveOn();
      }
      tries++;
    }, 100);
    var moveOn = function() { clearInterval(interval); nextTest(); };
  }
  
  function startFocusTest() {
    var plugin = document.getElementById('plugin');
    ok(plugin, "Got plugin element.");
    
    var condition = function() plugin.getBoundingClientRect().width == 400;
    waitForCondition(condition, afterPluginInsertion, "Waited too long for plugin to show up in page");
  }
    
  function afterPluginInsertion() {
    var plugin = document.getElementById('plugin');
    var objLoadingContent = plugin.QueryInterface(Ci.nsIObjectLoadingContent);
    ok(!objLoadingContent.activated, "plugin should not be activated");

    synthesizeMouseAtCenter(plugin, {}, window);
    var condition = function() objLoadingContent.activated;
    waitForCondition(condition, afterPluginActivation, "Waited too long for plugin to activate");
  }
    
  function afterPluginActivation() { 
    var plugin = document.getElementById('plugin');
    var objLoadingContent = plugin.QueryInterface(Ci.nsIObjectLoadingContent);
    ok(objLoadingContent.activated, "plugin should be activated now");
    is(plugin.getMouseUpEventCount(), 0, "Plugin should not have received mouse events yet.");

    synthesizeMouseAtCenter(plugin, {}, window);
    var condition = function() plugin.getMouseUpEventCount() > 0;
    waitForCondition(condition, afterFirstClick, "Waited too long for plugin to receive the mouse click");
  }
  
  function afterFirstClick() {
    var plugin = document.getElementById('plugin');
    is(plugin.getMouseUpEventCount(), 1, "Plugin should have received 1 mouse up event.");

    testsFinished();
  }

  function testsFinished() {
    try {
      SpecialPowers.clearUserPref(clickToPlayPref);
    }   
    catch(e) {
      ok(false, "Couldn't reset click-to-play pref");
    }
    SimpleTest.finish();
  }
    
  SpecialPowers.setBoolPref(clickToPlayPref, true);
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  
  document.write('<embed id="plugin" type="application/x-test" width="400" height="400" drawmode="solid" color="FF00FFFF"></embed>');
  SimpleTest.executeSoon(startFocusTest);

  </script>
</body>
</html>
